Move templates to separate folder
