+++ src/routes/mod.rs
@@ -1,0 +1,4 @@
+pub mod authors;
+pub mod books;
+pub mod public;
+pub mod series;
